07-building-chat-applications/chat-webapp/requirements.txt
flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
azure-identity==1.23.0
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.6.10
openai==1.95.1
httpx==0.28.1
python-dotenv==1.1.1
pydantic-settings==2.10.1
pydantic==2.11.7
python-json-logger==3.3.0
azure-cosmos==4.6.0
pytest==8.4.0
pytest-asyncio==1.0.0a1
pytest-cov==6.2.1
black==25.1.0
flake8==7.3.0